refactor(invoice): per-invoice REQUIRES_NEW finalize so one failure can't roll back the batch run
